ocr-api-py/pyproject.toml
[project]
name = "ocr-api"
version = "0.1.0"
description = "API for using OCR tools"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.1",
    "python-doctr[torch]>=0.9.0",
    "fastapi[standard]>=0.112.1",
    "uvicorn>=0.30.6",
    "surya-ocr>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
